Add enum to represent table sections
Switching on an enum is the accepted best practice for this sort of thing. It is clearer and easier to read and maintain than a series of if statements and makes it easier for others to contribute :)

This PR also adjusts the default table cell height on this screen from 50 to 44 to be consistent with the rest of the settings.